pyproject.toml
[project]
name = "testgrpc"
version = "0.1.0"
description = ""
authors = [
    {name = "colosss",email = "kocnevs.jk@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
]

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
